Cold Boot Process Implemented
The testing framework to leverage the AVR interpreter into Screeps
is in place and the interpreter boots. There is a routine to create
a VM for a creep. Since a creep program accepts certain particular
arguments at the moment.

I am aware that other people might want to write an AVR program
that starts up with no arguments. Essentially, a from scratch
start and I hope to implement that fairly soon. For now, this is
just how it happened to work out.

The `start` function takes two arguments. The first is a Room class
and the second is a Creep class. In a future version I will add
support for a argumentless start and one for a room.

The serialization and deserialization is in place but only partially
tested. There is not a 100% confirmation on the proof of concept
until I actually have a program persist across ticks.